+++ ui/components/designs/patterns/MesheryPatterns.tsx
@@ -60,12 +60,12 @@ import _ from 'lodash';
 import { getMeshModels } from '../../../api/meshmodel';
 import { modifyRJSFSchema } from '../../../utils/utils';
 import { Edit as EditIcon } from '@/assets/icons';
-import { updateVisibleColumns } from '../../../utils/responsive-column';
-import { useWindowDimensions } from '../../../utils/dimension';
-import InfoModal from '../../General/Modals/Information/InfoModal';
-import { InfoOutlined as InfoOutlinedIcon } from '@/assets/icons';
-import { DefaultTableCell, SortableTableCell } from '../../connections/common';
-import DefaultError from '../../General/error-404/index';
+import { updateVisibleColumns } from '../../utils/responsive-column';
+import { useWindowDimensions } from '../../utils/dimension';
+import InfoModal from '../shared/Modal/Information/InfoModal';
+import { InfoOutlined as InfoOutlinedIcon } from '@/components/icons';
+import { DefaultTableCell, SortableTableCell } from '../connections/common';
+import DefaultError from '../General/error-404/index';
 import CAN from '@/utils/can';
 import { keys } from '@/utils/permission_constants';
 import ExportModal from '../../ExportModal';
