+++ ui/components/filters/Filters.tsx
@@ -54,8 +54,8 @@ import {
 } from '@sistent/sistent';
 import { updateVisibleColumns } from '../../utils/responsive-column';
 import { useWindowDimensions } from '../../utils/dimension';
-import InfoModal from '../General/Modals/Information/InfoModal';
-import { InfoOutlined as InfoOutlinedIcon } from '@/assets/icons';
+import InfoModal from '../shared/Modal/Information/InfoModal';
+import { InfoOutlined as InfoOutlinedIcon } from '@/components/icons';
 import { DefaultTableCell, SortableTableCell } from '../connections/common/index';
 import CAN from '@/utils/can';
 import { keys } from '@/utils/permission_constants';
