+++ ui/components/shared/Modal/Information/styles.tsx
